Avoid specialized Translation types
This was a bit of an annoying step, as I had to split modules into smaller
modules to avoid circular dependencies.
