+++ ouroboros-consensus/src/Ouroboros/Consensus/HardFork/Combinator/Abstract/SingleEraBlock.hs
@@ -1,0 +1,56 @@
+{-# LANGUAGE FlexibleContexts #-}
+
+module Ouroboros.Consensus.HardFork.Combinator.Abstract.SingleEraBlock (
+    SingleEraBlock(..)
+  , singleEraTransition'
+  , proxySingle
+  ) where
+
+import           Cardano.Slotting.Slot
+import           Data.Proxy
+
+import           Ouroboros.Consensus.Block
+import           Ouroboros.Consensus.Ledger.Abstract
+import           Ouroboros.Consensus.Ledger.SupportsMempool
+import           Ouroboros.Consensus.Ledger.SupportsProtocol
+
+import           Ouroboros.Consensus.HardFork.Combinator.Info
+import           Ouroboros.Consensus.HardFork.Combinator.PartialConfig
+
+{-------------------------------------------------------------------------------
+  SingleEraBlock
+-------------------------------------------------------------------------------}
+
+-- | Blocks from which we can assemble a hard fork
+class ( LedgerSupportsProtocol blk
+      , LedgerSupportsMempool blk
+      , HasTxId (GenTx blk)
+      , QueryLedger blk
+      , CanForge blk
+      , HasPartialConsensusConfig (BlockProtocol blk)
+      , HasPartialLedgerConfig blk
+      , ConvertRawHash blk
+        -- Instances required to support testing
+      , Show blk
+      , Show (Header blk)
+      ) => SingleEraBlock blk where
+
+  -- | Era transition
+  --
+  -- This should only report the transition point once it is stable (rollback
+  -- cannot affect it anymore).
+  --
+  -- This takes the partial config rather than the full config for the same
+  -- reason as 'singleEraParam'.
+  singleEraTransition :: PartialLedgerConfig blk -> LedgerState blk -> Maybe EpochNo
+
+  -- | Era information (for use in error messages)
+  singleEraInfo       :: proxy blk -> SingleEraInfo blk
+
+proxySingle :: Proxy SingleEraBlock
+proxySingle = Proxy
+
+singleEraTransition' :: SingleEraBlock blk
+                     => WrapPartialLedgerConfig blk
+                     -> LedgerState blk -> Maybe EpochNo
+singleEraTransition' = singleEraTransition . unwrapPartialLedgerConfig
