+++ ouroboros-consensus/src/Ouroboros/Consensus/HardFork/Combinator/Protocol/LedgerView.hs
@@ -19,9 +19,9 @@ import           Ouroboros.Consensus.TypeFamilyWrappers
 import           Ouroboros.Consensus.HardFork.Combinator.Abstract
 import           Ouroboros.Consensus.HardFork.Combinator.Basics
 import           Ouroboros.Consensus.HardFork.Combinator.PartialConfig
-import           Ouroboros.Consensus.HardFork.Combinator.State.Infra
-                     (HardForkState_, TransitionOrTip (..))
 import qualified Ouroboros.Consensus.HardFork.Combinator.State.Infra as State
+import           Ouroboros.Consensus.HardFork.Combinator.State.Types
+                     (HardForkState_, TransitionOrTip (..))
 
 {-------------------------------------------------------------------------------
   Ledger view for single era
