+++ ouroboros-consensus/src/Ouroboros/Consensus/HardFork/Combinator/State/Infra.hs
@@ -1,32 +1,16 @@
-{-# LANGUAGE DataKinds             #-}
-{-# LANGUAGE DeriveAnyClass        #-}
-{-# LANGUAGE DeriveGeneric         #-}
-{-# LANGUAGE DerivingVia           #-}
-{-# LANGUAGE EmptyCase             #-}
-{-# LANGUAGE FlexibleContexts      #-}
-{-# LANGUAGE FlexibleInstances     #-}
-{-# LANGUAGE LambdaCase            #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE QuantifiedConstraints #-}
-{-# LANGUAGE RankNTypes            #-}
-{-# LANGUAGE RecordWildCards       #-}
-{-# LANGUAGE ScopedTypeVariables   #-}
-{-# LANGUAGE StandaloneDeriving    #-}
-{-# LANGUAGE TypeApplications      #-}
-{-# LANGUAGE TypeFamilies          #-}
-{-# LANGUAGE TypeOperators         #-}
+{-# LANGUAGE DataKinds           #-}
+{-# LANGUAGE EmptyCase           #-}
+{-# LANGUAGE FlexibleContexts    #-}
+{-# LANGUAGE LambdaCase          #-}
+{-# LANGUAGE RankNTypes          #-}
+{-# LANGUAGE RecordWildCards     #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeFamilies        #-}
+{-# LANGUAGE TypeOperators       #-}
 
--- | 'HardForkState' infrastructure that is independent from @.Basics@
---
--- Separate module to avoid circular module dependencies
 module Ouroboros.Consensus.HardFork.Combinator.State.Infra (
-    -- * Types
-    HardForkState_(..)
-  , HardForkState
-  , initHardForkState
-  , Past(..)
-  , Snapshot(..)
-  , Current(..)
+    -- * Initialization
+    initHardForkState
     -- * GC
   , tickAllPast
     -- * Lifting 'Telescope' operations
@@ -36,26 +20,19 @@ module Ouroboros.Consensus.HardFork.Combinator.State.Infra (
   , bihczipWith
   , fromTZ
     -- * Aligning
-  , Translate(..)
   , align
     -- * Rewinding
   , retractToSlot
     -- * EpochInfo/Summary
-  , TransitionOrTip(..)
   , transitionOrTip
   , reconstructSummary
   ) where
 
 import           Prelude hiding (sequence)
 
-import           Codec.Serialise
-import           Data.Functor.Identity
 import           Data.Functor.Product
 import           Data.SOP.Strict hiding (shape)
-import           Data.Word
-import           GHC.Generics (Generic)
 
-import           Cardano.Prelude (NoUnexpectedThunks)
 import           Cardano.Slotting.Slot
 
 import           Ouroboros.Consensus.Config.SecurityParam
@@ -64,11 +41,11 @@ import           Ouroboros.Consensus.HardFork.History (Bound (..), EraEnd (..),
 import qualified Ouroboros.Consensus.HardFork.History as History
 import           Ouroboros.Consensus.Ledger.Abstract
 import           Ouroboros.Consensus.Util.Counting
-import           Ouroboros.Consensus.Util.SOP
 
-import           Ouroboros.Consensus.HardFork.Combinator.Abstract
+import           Ouroboros.Consensus.HardFork.Combinator.Abstract.SingleEraBlock
 import           Ouroboros.Consensus.HardFork.Combinator.PartialConfig
-import           Ouroboros.Consensus.HardFork.Combinator.Util.DerivingVia
+import           Ouroboros.Consensus.HardFork.Combinator.State.Lift
+import           Ouroboros.Consensus.HardFork.Combinator.State.Types
 import           Ouroboros.Consensus.HardFork.Combinator.Util.InPairs (InPairs,
                      Requiring (..))
 import qualified Ouroboros.Consensus.HardFork.Combinator.Util.InPairs as InPairs
@@ -80,83 +57,6 @@ import           Ouroboros.Consensus.HardFork.Combinator.Util.Telescope
 import qualified Ouroboros.Consensus.HardFork.Combinator.Util.Telescope as Telescope
 
 {-------------------------------------------------------------------------------
-  Types
--------------------------------------------------------------------------------}
-
--- | Generic hard fork state
---
--- This is used both for the consensus state and the ledger state.
-newtype HardForkState_ g f xs = HardForkState {
-      getHardForkState :: Telescope (Past g) (Current f) xs
-    }
-
--- | Hard for state where both 'Past' and 'Current' use the same functor
---
--- In most cases this is what we need; we only end up with different functors
--- after things like 'match'.
-type HardForkState f = HardForkState_ f f
-
--- | Information about the current era
-data Current f blk = Current {
-      currentStart :: !Bound
-    , currentState :: !(f blk)
-    }
-  deriving (Generic)
-
--- | Information about a past era
-data Past f blk = Past {
-      pastStart    :: !Bound
-    , pastEnd      :: !Bound
-    , pastSnapshot :: !(Snapshot f blk)
-    }
-  deriving (Generic)
-
--- | Past snapshot
---
--- We record for each past era how many blocks have been applied to /any/
--- subsequent era. Here is an example with @k = 3@ with three ledgers
--- @A@, @B@ and @C@, with maximum roll back marked for a few states:
---
--- > Initial ledger   Curr A0
--- >
--- > Apply block      Curr A1                      <--\
--- >                                                  |
--- > Transition       Past 0 A1, Curr B0              |
--- > Apply block      Past 1 A1, Curr B1              |  <--\
--- >                                                  |     |
--- > Apply block      Past 2 A1, Curr B2              |     |
--- >                                                  |     |
--- > Transition       Past 2 A1, Past 0 B2, Curr C0   |     |
--- > Apply block      Past 3 A1, Past 1 B2, Curr C1   /     |  <--\
--- >                                                        |     |
--- > Apply block      Past 4 A1, Past 2 B2, Curr C2         |     |
--- > GC               Past GCd,  Past 2 B2, Curr C2         /     |
--- >                                                              |
--- > Apply block      Past GCd,  Past 3 B2, Curr C3               |
--- >                                                              |
--- > Apply block      Past GCd,  Past 4 B2, Curr C4               |
--- > GC               Past GCd,  Past GCd,  Curr C4               /
---
--- Note that at the point where past states are GCed, we indeed can no longer
--- roll back to the point before the corresponding transitions.
-data Snapshot f blk =
-    -- | Past snapshot still available
-    --
-    -- Invariant: the count must be @<= k@ (see diagram above).
-    Snapshot !Word64 !(f blk)
-
-    -- | Past consensus state not available anymore
-    --
-    -- After @k@ blocks have been applied, we are sure that we don't need
-    -- the old consensus state anymore and so we don't need to keep it around.
-  | NoSnapshot
-  deriving (Generic)
-
-getSnapshot :: Snapshot f blk -> Maybe (f blk)
-getSnapshot (Snapshot _ st) = Just st
-getSnapshot NoSnapshot      = Nothing
-
-{-------------------------------------------------------------------------------
   Initialization
 -------------------------------------------------------------------------------}
 
@@ -165,53 +65,6 @@ initHardForkState st = HardForkState $ TZ $ Current {
       currentStart = History.initBound
     , currentState = st
     }
-
-{-------------------------------------------------------------------------------
-  Lifting functions on @f@ to @Current @f@
--------------------------------------------------------------------------------}
-
-lift :: (f blk -> f' blk) -> Current f blk -> Current f' blk
-lift f = runIdentity . liftM (Identity . f)
-
-liftM :: Functor m
-      => (f blk -> m (f' blk)) -> Current f blk -> m (Current f' blk)
-liftM f (Current start cur) = Current start <$> f cur
-
-{-------------------------------------------------------------------------------
-  Lifting functions on @f@ to @Past f@
--------------------------------------------------------------------------------}
-
-liftPast :: (f blk -> f' blk) -> Past f blk -> Past f' blk
-liftPast f = runIdentity . liftPastM (Identity . f)
-
-liftPastM :: Applicative m
-          => (f blk -> m (f' blk)) -> Past f blk -> m (Past f' blk)
-liftPastM f (Past start end snapshot) =
-    Past start end <$>
-      case snapshot of
-        NoSnapshot    -> pure NoSnapshot
-        Snapshot n st -> Snapshot n <$> f st
-
-{-------------------------------------------------------------------------------
-  SOP class instances
-
-  These are convenient, allowing us to treat the 'HardForkState' just like any
-  other SOP type; in particular, they deal with lifting functions to 'Current'.
--------------------------------------------------------------------------------}
-
-type instance Prod    (HardForkState_ g)   = NP
-type instance SListIN (HardForkState_ g)   = SListI
-type instance AllN    (HardForkState_ g) c = All c
-
-instance HAp (HardForkState_ g) where
-  hap np (HardForkState st) = HardForkState $
-      hap (map_NP' (Fn . lift . apFn) np) st
-
-instance HSequence (HardForkState_ g) where
-  hctraverse' = \p f (HardForkState st) -> HardForkState <$>
-                                              hctraverse' p (liftM f) st
-  htraverse' = hctraverse' (Proxy @Top)
-  hsequence' = htraverse' unComp
 
 {-------------------------------------------------------------------------------
   GC
@@ -260,7 +113,7 @@ sequence = \(HardForkState st) -> HardForkState <$>
     distrib (Current start st) = Comp $
         Current start <$> unComp st
 
-bihczipWith :: forall xs h g' g f' f. CanHardFork xs
+bihczipWith :: forall xs h g' g f' f. All SingleEraBlock xs
             => (forall blk. SingleEraBlock blk => h blk -> g blk -> g' blk)
             -> (forall blk. SingleEraBlock blk => h blk -> f blk -> f' blk)
             -> NP h xs -> HardForkState_ g f xs -> HardForkState_ g' f' xs
@@ -274,11 +127,7 @@ fromTZ = currentState . Telescope.fromTZ . getHardForkState
   Aligning
 -------------------------------------------------------------------------------}
 
-newtype Translate f x y = Translate {
-      translateWith :: EpochNo -> f x -> f y
-    }
-
-align :: forall xs h f f' f''. CanHardFork xs
+align :: forall xs h f f' f''. All SingleEraBlock xs
       => InPairs (Translate f) xs
       -> NP (f' -.-> f -.-> f'') xs
       -> HardForkState_ h f'  xs -- ^ State we are aligning with
@@ -347,19 +196,13 @@ retractToSlot slot (HardForkState st) =
     retract = Retract $ \past _oldCur ->
         Current (pastStart past) <$> getSnapshot (pastSnapshot past)
 
+getSnapshot :: Snapshot f blk -> Maybe (f blk)
+getSnapshot (Snapshot _ st) = Just st
+getSnapshot NoSnapshot      = Nothing
+
 {-------------------------------------------------------------------------------
   Summary/EpochInfo
 -------------------------------------------------------------------------------}
-
--- | Property of a particular ledger state: transition to the next era if known,
--- or the tip of the ledger otherwise.
-data TransitionOrTip =
-    -- | Transition to the next era has been confirmed and is stable
-    TransitionAt !(WithOrigin SlotNo) !EpochNo
-
-    -- | Transition to the next era not yet known; we reported ledger tip
-  | LedgerTip !(WithOrigin SlotNo)
-  deriving (Show)
 
 transitionOrTip :: SingleEraBlock blk
                 => WrapPartialLedgerConfig blk
@@ -370,7 +213,7 @@ transitionOrTip cfg st =
       Just epoch -> TransitionAt (ledgerTipSlot st) epoch
       Nothing    -> LedgerTip (ledgerTipSlot st)
 
-reconstructSummary :: forall g f xs. CanHardFork xs
+reconstructSummary :: forall g f xs. All SingleEraBlock xs
                    => History.Shape xs
                    -> NP (f -.-> K TransitionOrTip) xs
                    -- ^ Return the 'EpochNo' of the transition to the next
@@ -438,47 +281,3 @@ reconstructSummary (History.Shape shape) transition (HardForkState st) =
         . History.slotToEpochBound params start
         . History.addSlots (History.safeFromTip eraSafeZone)
         . fromWithOrigin (boundSlot start)
-
-{-------------------------------------------------------------------------------
-  Instances
--------------------------------------------------------------------------------}
-
-deriving instance Eq                 (f blk) => Eq                 (Current f blk)
-deriving instance Show               (f blk) => Show               (Current f blk)
-deriving instance NoUnexpectedThunks (f blk) => NoUnexpectedThunks (Current f blk)
-
-deriving instance Eq                 (f blk) => Eq                 (Past f blk)
-deriving instance Show               (f blk) => Show               (Past f blk)
-deriving instance NoUnexpectedThunks (f blk) => NoUnexpectedThunks (Past f blk)
-
-deriving instance Eq                 (f blk) => Eq                 (Snapshot f blk)
-deriving instance Show               (f blk) => Show               (Snapshot f blk)
-deriving instance NoUnexpectedThunks (f blk) => NoUnexpectedThunks (Snapshot f blk)
-
-deriving via LiftTelescope (Past g) (Current f) xs
-         instance ( CanHardFork xs
-                  , forall blk. SingleEraBlock blk => Show (f blk)
-                  , forall blk. SingleEraBlock blk => Show (g blk)
-                  ) => Show (HardForkState_ g f xs)
-
-deriving via LiftTelescope (Past g) (Current f) xs
-         instance ( CanHardFork xs
-                  , forall blk. SingleEraBlock blk => Eq (f blk)
-                  , forall blk. SingleEraBlock blk => Eq (g blk)
-                  ) => Eq (HardForkState_ g f xs)
-
-deriving via LiftNamedTelescope "HardForkState" (Past g) (Current f) xs
-         instance ( CanHardFork xs
-                  , forall blk. SingleEraBlock blk => NoUnexpectedThunks (f blk)
-                  , forall blk. SingleEraBlock blk => NoUnexpectedThunks (g blk)
-                  ) => NoUnexpectedThunks (HardForkState_ g f xs)
-
-{-------------------------------------------------------------------------------
-  Serialisation
-
-  This is primarily useful for tests.
--------------------------------------------------------------------------------}
-
-deriving instance Serialise (f blk) => Serialise (Current  f blk)
-deriving instance Serialise (f blk) => Serialise (Past     f blk)
-deriving instance Serialise (f blk) => Serialise (Snapshot f blk)
